packet: switch to protobuf implementation
