+++ test/packet-decrypt-header-test.c
@@ -31,25 +31,21 @@
 #include "tracing.h"
 
 int main(void) {
-	//decrypted header buffers
-	buffer_t *decrypted_header = buffer_create_on_heap(255, 255);
-	buffer_t *decrypted_message_nonce = buffer_create_on_heap(crypto_secretbox_NONCEBYTES, crypto_secretbox_NONCEBYTES);
+	return_status status = return_status_init();
 
 	//generate keys
-	buffer_t *header_key = buffer_create_on_heap(HEADER_KEY_SIZE, HEADER_KEY_SIZE);
+	buffer_t *header_key = buffer_create_on_heap(crypto_aead_chacha20poly1305_KEYBYTES, crypto_aead_chacha20poly1305_KEYBYTES);
 	buffer_t *message_key = buffer_create_on_heap(crypto_secretbox_KEYBYTES, crypto_secretbox_KEYBYTES);
 	buffer_t *public_identity_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *public_ephemeral_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *public_prekey = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
-	buffer_t *extracted_public_identity_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
-	buffer_t *extracted_public_ephemeral_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
-	buffer_t *extracted_public_prekey = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 
 	buffer_t *header = buffer_create_on_heap(4, 4);
 	buffer_create_from_string(message, "Hello world!\n");
-	buffer_t *packet = buffer_create_on_heap(3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE, 3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE);
 
-	return_status status = return_status_init();
+	buffer_t *packet = NULL;
+	buffer_t *decrypted_header = NULL;
+
 
 	if(sodium_init() == -1) {
 		throw(INIT_ERROR, "Failed to initialize libsodium.");
@@ -60,26 +56,20 @@ int main(void) {
 	header->content[1] = 0x02;
 	header->content[2] = 0x03;
 	header->content[3] = 0x04;
-	unsigned char packet_type = 1;
+	molch_message_type packet_type = 1;
 	printf("Packet type: %02x\n", packet_type);
-	const unsigned char current_protocol_version = 2;
-	printf("Current protocol version: %02x\n", current_protocol_version);
-	const unsigned char highest_supported_protocol_version = 3;
-	printf("Highest supported protocol version: %02x\n", highest_supported_protocol_version);
 	putchar('\n');
 
 	//NORMAL MESSAGE
 	printf("NORMAL MESSAGE\n");
 	int status_int = 0;
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			NULL,
 			NULL,
 			NULL);
@@ -87,13 +77,9 @@ int main(void) {
 
 	//now decrypt the header
 	status = packet_decrypt_header(
+			&decrypted_header,
 			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			NULL,
-			NULL,
-			NULL);
+			header_key);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt the header.");
 
 
@@ -101,10 +87,6 @@ int main(void) {
 		throw(INVALID_VALUE, "Decrypted header isn't of the same length.");
 	}
 	printf("Decrypted header has the same length.\n\n");
-
-	printf("Decrypted message nonce (%zu Bytes):\n", decrypted_message_nonce->content_length);
-	print_hex(decrypted_message_nonce);
-	putchar('\n');
 
 	//compare headers
 	if (buffer_compare(header, decrypted_header) != 0) {
@@ -116,13 +98,9 @@ int main(void) {
 	printf("Manipulating header length.\n");
 	packet->content[2]++;
 	status = packet_decrypt_header(
+			&decrypted_header,
 			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			NULL,
-			NULL,
-			NULL);
+			header_key);
 	if (status.status == SUCCESS) {
 		throw(GENERIC_ERROR, "Manipulated packet was accepted.");
 	} else {
@@ -135,15 +113,11 @@ int main(void) {
 	packet->content[2]--;
 	//check if it decrypts manipulated packets (manipulated header)
 	printf("Manipulate header.\n");
-	packet->content[3 + HEADER_NONCE_SIZE + 1] ^= 0x12;
+	packet->content[3 + crypto_aead_chacha20poly1305_NPUBBYTES + 1] ^= 0x12;
 	status = packet_decrypt_header(
+			&decrypted_header,
 			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			NULL,
-			NULL,
-			NULL);
+			header_key);
 	if (status.status == SUCCESS) {
 		throw(GENERIC_ERROR, "Manipulated packet was accepted.");
 	} else {
@@ -153,7 +127,7 @@ int main(void) {
 	printf("Header manipulation detected!\n\n");
 
 	//undo header manipulation
-	packet->content[3 + HEADER_NONCE_SIZE + 1] ^= 0x12;
+	packet->content[3 + crypto_aead_chacha20poly1305_NPUBBYTES + 1] ^= 0x12;
 
 	//PREKEY MESSAGE
 	printf("PREKEY_MESSAGE\n");
@@ -171,17 +145,17 @@ int main(void) {
 		throw(KEYGENERATION_FAILED, "Failed to generate public prekey.");
 	}
 
-	buffer_clear(packet);
+	buffer_destroy_from_heap(packet);
+	packet = NULL;
+
 	packet_type = PREKEY_MESSAGE;
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			public_identity_key,
 			public_ephemeral_key,
 			public_prekey);
@@ -189,13 +163,9 @@ int main(void) {
 
 	//now decrypt the header
 	status = packet_decrypt_header(
+			&decrypted_header,
 			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			extracted_public_identity_key,
-			extracted_public_ephemeral_key,
-			extracted_public_prekey);
+			header_key);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt the header.");
 
 	if (decrypted_header->content_length != header->content_length) {
@@ -203,45 +173,25 @@ int main(void) {
 	}
 	printf("Decrypted header has the same length.\n\n");
 
-	printf("Decrypted message nonce (%zu Bytes):\n", decrypted_message_nonce->content_length);
-	print_hex(decrypted_message_nonce);
-	putchar('\n');
-
 	//compare headers
 	if (buffer_compare(header, decrypted_header) != 0) {
 		throw(INVALID_VALUE, "Decrypted header doesn't match.");
 	}
 	printf("Decrypted header matches.\n");
 
-	//compare public keys
-	if (buffer_compare(public_identity_key, extracted_public_identity_key) != 0) {
-		throw(INVALID_VALUE, "Extracted public identity key doesn't match.");
-	}
-	printf("Extracted public identity key matches!\n");
-
-	if (buffer_compare(public_ephemeral_key, extracted_public_ephemeral_key) != 0) {
-		throw(INVALID_VALUE, "Extracted public ephemeral key doesn't match.");
-	}
-	printf("Extracted public ephemeral key matches!\n");
-
-	if (buffer_compare(public_prekey, extracted_public_prekey) != 0) {
-		throw(INVALID_VALUE, "Extracted public prekey doesn't match.");
-	}
-	printf("Extracted public prekey matches!\n");
-
 cleanup:
-	buffer_destroy_from_heap(decrypted_header);
-	buffer_destroy_from_heap(decrypted_message_nonce);
 	buffer_destroy_from_heap(header_key);
 	buffer_destroy_from_heap(message_key);
 	buffer_destroy_from_heap(header);
-	buffer_destroy_from_heap(packet);
 	buffer_destroy_from_heap(public_identity_key);
 	buffer_destroy_from_heap(public_ephemeral_key);
 	buffer_destroy_from_heap(public_prekey);
-	buffer_destroy_from_heap(extracted_public_identity_key);
-	buffer_destroy_from_heap(extracted_public_ephemeral_key);
-	buffer_destroy_from_heap(extracted_public_prekey);
+	if (packet != NULL) {
+		buffer_destroy_from_heap(packet);
+	}
+	if (decrypted_header != NULL) {
+		buffer_destroy_from_heap(decrypted_header);
+	}
 
 	if (status.status != SUCCESS) {
 		print_errors(&status);
