+++ test/packet-decrypt-message-test.c
@@ -33,16 +33,15 @@
 int main(void) {
 	buffer_create_from_string(message, "Hello world!\n");
 	//create buffers
-	buffer_t *header_key = buffer_create_on_heap(HEADER_KEY_SIZE, HEADER_KEY_SIZE);
+	buffer_t *header_key = buffer_create_on_heap(crypto_aead_chacha20poly1305_KEYBYTES, crypto_aead_chacha20poly1305_KEYBYTES);
 	buffer_t *message_key = buffer_create_on_heap(crypto_secretbox_KEYBYTES, crypto_secretbox_KEYBYTES);
 	buffer_t *public_identity_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *public_ephemeral_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *public_prekey = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *header = buffer_create_on_heap(4, 4);
-	buffer_t *packet = buffer_create_on_heap(3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE, 3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE);
-	buffer_t *decrypted_header = buffer_create_on_heap(255, 255);
-	buffer_t *decrypted_message_nonce = buffer_create_on_heap(crypto_secretbox_NONCEBYTES, crypto_secretbox_NONCEBYTES);
-	buffer_t *decrypted_message = buffer_create_on_heap(packet->content_length, packet->content_length);
+
+	buffer_t *packet = NULL;
+	buffer_t *decrypted_message = NULL;
 
 	return_status status = return_status_init();
 
@@ -55,52 +54,29 @@ int main(void) {
 	header->content[1] = 0x02;
 	header->content[2] = 0x03;
 	header->content[3] = 0x04;
-	unsigned char packet_type = 1;
+	molch_message_type packet_type = NORMAL_MESSAGE;
 	printf("Packet type: %02x\n", packet_type);
-	const unsigned char current_protocol_version = 2;
-	printf("Current protocol version: %02x\n", current_protocol_version);
-	const unsigned char highest_supported_protocol_version = 3;
-	printf("Highest supported protocol version: %02x\n", highest_supported_protocol_version);
 	putchar('\n');
 
 	//NORMAL MESSAGE
 	printf("NORMAL MESSAGE\n");
 	int status_int = 0;
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			NULL,
 			NULL,
 			NULL);
 	throw_on_error(GENERIC_ERROR, "Failed to create and print message.");
 
-	//now decrypt the header
-	status = packet_decrypt_header(
-			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			NULL,
-			NULL,
-			NULL);
-	buffer_clear(decrypted_header);
-	throw_on_error(DECRYPT_ERROR, "Failed to decrypt header.");
-
-	printf("Decrypted message nonce (%zu Bytes):\n", decrypted_message_nonce->content_length);
-	print_hex(decrypted_message_nonce);
-	putchar('\n');
-
 	//now decrypt the message
 	status = packet_decrypt_message(
+			&decrypted_message,
 			packet,
-			decrypted_message,
-			decrypted_message_nonce,
 			message_key);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt message.");
 
@@ -114,18 +90,19 @@ int main(void) {
 	if (buffer_compare(message, decrypted_message) != 0) {
 		throw(INVALID_VALUE, "Decrypted message doesn't match.");
 	}
-	buffer_clear(decrypted_message);
 	printf("Decrypted message is the same.\n\n");
 
 	//manipulate the message
 	packet->content[packet->content_length - crypto_secretbox_MACBYTES - 1] ^= 0xf0;
 	printf("Manipulating message.\n");
 
+	buffer_destroy_from_heap(decrypted_message);
+	decrypted_message = NULL;
+
 	//try to decrypt
 	status = packet_decrypt_message(
+			&decrypted_message,
 			packet,
-			decrypted_message,
-			decrypted_message_nonce,
 			message_key);
 	if (status.status == SUCCESS) { //message was decrypted although it shouldn't
 		throw(GENERIC_ERROR, "Decrypted manipulated message.");
@@ -150,42 +127,26 @@ int main(void) {
 		throw(KEYGENERATION_FAILED, "Failed to generate public prekey.");
 	}
 
-	buffer_clear(packet);
+	buffer_destroy_from_heap(packet);
+	packet = NULL;
+
 	packet_type = PREKEY_MESSAGE;
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			public_identity_key,
 			public_ephemeral_key,
 			public_prekey);
 	throw_on_error(GENERIC_ERROR, "Failed to create and print message.");
 
-	//now decrypt the header
-	status = packet_decrypt_header(
-			packet,
-			decrypted_header,
-			decrypted_message_nonce,
-			header_key,
-			NULL,
-			NULL,
-			NULL);
-	throw_on_error(DECRYPT_ERROR, "Failed to decrypt header.");
-
-	printf("Decrypted message nonce (%zu Bytes):\n", decrypted_message_nonce->content_length);
-	print_hex(decrypted_message_nonce);
-	putchar('\n');
-
 	//now decrypt the message
 	status = packet_decrypt_message(
+			&decrypted_message,
 			packet,
-			decrypted_message,
-			decrypted_message_nonce,
 			message_key);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt message.");
 
@@ -199,17 +160,18 @@ int main(void) {
 	if (buffer_compare(message, decrypted_message) != 0) {
 		throw(INVALID_VALUE, "Decrypted message doesn't match.");
 	}
-	buffer_clear(decrypted_message);
 	printf("Decrypted message is the same.\n");
 
 cleanup:
 	buffer_destroy_from_heap(header_key);
 	buffer_destroy_from_heap(message_key);
 	buffer_destroy_from_heap(header);
-	buffer_destroy_from_heap(packet);
-	buffer_destroy_from_heap(decrypted_header);
-	buffer_destroy_from_heap(decrypted_message_nonce);
-	buffer_destroy_from_heap(decrypted_message);
+	if (packet != NULL) {
+		buffer_destroy_from_heap(packet);
+	}
+	if (decrypted_message != NULL) {
+		buffer_destroy_from_heap(decrypted_message);
+	}
 	buffer_destroy_from_heap(public_identity_key);
 	buffer_destroy_from_heap(public_ephemeral_key);
 	buffer_destroy_from_heap(public_prekey);
