+++ test/packet-decrypt-test.c
@@ -35,7 +35,7 @@ int main(void) {
 	buffer_create_from_string(message, "Hello world!\n");
 
 	//create buffers
-	buffer_t *header_key = buffer_create_on_heap(HEADER_KEY_SIZE, HEADER_KEY_SIZE);
+	buffer_t *header_key = buffer_create_on_heap(crypto_aead_chacha20poly1305_KEYBYTES, crypto_aead_chacha20poly1305_KEYBYTES);
 	buffer_t *message_key = buffer_create_on_heap(crypto_secretbox_KEYBYTES, crypto_secretbox_KEYBYTES);
 	buffer_t *public_identity_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *public_ephemeral_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
@@ -44,9 +44,9 @@ int main(void) {
 	buffer_t *extracted_public_ephemeral_key = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *extracted_public_prekey = buffer_create_on_heap(PUBLIC_KEY_SIZE, PUBLIC_KEY_SIZE);
 	buffer_t *header = buffer_create_on_heap(4, 4);
-	buffer_t *packet = buffer_create_on_heap(3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE, 3 + HEADER_NONCE_SIZE + crypto_secretbox_MACBYTES + crypto_secretbox_NONCEBYTES + message->content_length + header->content_length + crypto_secretbox_MACBYTES + 255 + 3 * PUBLIC_KEY_SIZE);
-	buffer_t *decrypted_header = buffer_create_on_heap(255, 255);
-	buffer_t *decrypted_message = buffer_create_on_heap(packet->content_length, packet->content_length);
+	buffer_t *packet = NULL;
+	buffer_t *decrypted_header = NULL;
+	buffer_t *decrypted_message = NULL;
 
 	return_status status = return_status_init();
 
@@ -59,52 +59,46 @@ int main(void) {
 	header->content[1] = 0x02;
 	header->content[2] = 0x03;
 	header->content[3] = 0x04;
-	unsigned char packet_type = NORMAL_MESSAGE;
+	molch_message_type packet_type = NORMAL_MESSAGE;
 	printf("Packet type: %02x\n", packet_type);
-	const unsigned char current_protocol_version = 2;
-	printf("Current protocol version: %02x\n", current_protocol_version);
-	const unsigned char highest_supported_protocol_version = 3;
-	printf("Highest supported protocol version: %02x\n", highest_supported_protocol_version);
 	putchar('\n');
 
 	//NORMAL MESSAGE
 	printf("NORMAL MESSAGE\n");
 	int status_int = 0;
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			NULL,
 			NULL,
 			NULL);
 	throw_on_error(CREATION_ERROR, "Failed to create and print normal message.");
 
 	//now decrypt the packet
-	unsigned char authenticated_packet_type;
-	unsigned char authenticated_current_protocol_version;
-	unsigned char authenticated_highest_supported_protocol_version;
+	molch_message_type extracted_packet_type;
+	uint32_t extracted_current_protocol_version;
+	uint32_t extracted_highest_supported_protocol_version;
 	status = packet_decrypt(
+			&extracted_current_protocol_version,
+			&extracted_highest_supported_protocol_version,
+			&extracted_packet_type,
+			&decrypted_header,
+			&decrypted_message,
 			packet,
-			&authenticated_packet_type,
-			&authenticated_current_protocol_version,
-			&authenticated_highest_supported_protocol_version,
-			decrypted_header,
 			header_key,
-			decrypted_message,
 			message_key,
 			NULL,
 			NULL,
 			NULL);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt the packet.");
 
-	if ((packet_type != authenticated_packet_type)
-		|| (current_protocol_version != authenticated_current_protocol_version)
-		|| (highest_supported_protocol_version != authenticated_highest_supported_protocol_version)) {
+	if ((packet_type != extracted_packet_type)
+		|| (extracted_current_protocol_version != 0)
+		|| (extracted_highest_supported_protocol_version != 0)) {
 		throw(DATA_FETCH_ERROR, "Failed to retrieve metadata.");
 	}
 
@@ -147,17 +141,22 @@ int main(void) {
 		throw(KEYGENERATION_FAILED, "Failed to generate public prekey.");
 	}
 
-	buffer_clear(packet);
+	buffer_destroy_from_heap(decrypted_header);
+	decrypted_header = NULL;
+	buffer_destroy_from_heap(decrypted_message);
+	decrypted_message = NULL;
+	buffer_destroy_from_heap(packet);
+	packet = NULL;
+
 	packet_type = PREKEY_MESSAGE;
+
 	status = create_and_print_message(
-			packet,
-			packet_type,
-			current_protocol_version,
-			highest_supported_protocol_version,
-			message,
-			message_key,
-			header,
+			&packet,
 			header_key,
+			message_key,
+			packet_type,
+			header,
+			message,
 			public_identity_key,
 			public_ephemeral_key,
 			public_prekey);
@@ -165,22 +164,22 @@ int main(void) {
 
 	//now decrypt the packet
 	status = packet_decrypt(
+			&extracted_current_protocol_version,
+			&extracted_highest_supported_protocol_version,
+			&extracted_packet_type,
+			&decrypted_header,
+			&decrypted_message,
 			packet,
-			&authenticated_packet_type,
-			&authenticated_current_protocol_version,
-			&authenticated_highest_supported_protocol_version,
-			decrypted_header,
 			header_key,
-			decrypted_message,
 			message_key,
 			extracted_public_identity_key,
 			extracted_public_ephemeral_key,
 			extracted_public_prekey);
 	throw_on_error(DECRYPT_ERROR, "Failed to decrypt the packet.");
 
-	if ((packet_type != authenticated_packet_type)
-			|| (current_protocol_version != authenticated_current_protocol_version)
-			|| (highest_supported_protocol_version != authenticated_highest_supported_protocol_version)) {
+	if ((packet_type != extracted_packet_type)
+			|| (extracted_current_protocol_version != 0)
+			|| (extracted_highest_supported_protocol_version != 0)) {
 		throw(DATA_FETCH_ERROR, "Failed to retrieve metadata.");
 	}
 
@@ -226,9 +225,15 @@ cleanup:
 	buffer_destroy_from_heap(header_key);
 	buffer_destroy_from_heap(message_key);
 	buffer_destroy_from_heap(header);
-	buffer_destroy_from_heap(packet);
-	buffer_destroy_from_heap(decrypted_header);
-	buffer_destroy_from_heap(decrypted_message);
+	if (packet != NULL) {
+		buffer_destroy_from_heap(packet);
+	}
+	if (decrypted_header != NULL) {
+		buffer_destroy_from_heap(decrypted_header);
+	}
+	if (decrypted_message != NULL) {
+		buffer_destroy_from_heap(decrypted_message);
+	}
 	buffer_destroy_from_heap(public_identity_key);
 	buffer_destroy_from_heap(public_ephemeral_key);
 	buffer_destroy_from_heap(public_prekey);
